pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foodbot"
version = "0.1.0"
description = "Telegram-first FoodBOT onboarding and profile service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.4,<4.0",
  "SQLAlchemy[asyncio]>=2.0.28,<3.0",
  "asyncpg>=0.29.0,<1.0",
  "alembic>=1.13.1,<2.0",
  "pydantic>=2.6.4,<3.0",
  "python-dotenv>=1.0.1,<2.0",
  "sqlparse>=0.5.0,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.1,<9.0",
  "pytest-asyncio>=0.23.5,<1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools]
packages = ["bot"]

[tool.setuptools.package-data]
bot = ["py.typed"]
